requirements.txt
pandas
jinja2
weasyprint
aiogram==3.13.1
python-dotenv
Pillow

